Fixed bug in freq calc, saving functions
Fixed bug in freq calc where if the cut off frequency just happened to
be exactly equal in both the calibrator and on the phantom, it would set
the max freq to 1000.

Fixed bug in saving functions so now it outputs a SUM file and saves the
fits.

Still existing bug: It won't extract the right diode name per file, have
to manually set still.